Make the paraview gui use the main python intrepreter
Instead of creating a new python intrepreter, use the main python
intrepreter in the paraview gui.  Other interpreters, owned by objects
like the python programmable filter will still be created as new
interpreters and function the same as before.

By having the gui use the main interpreter instead of a new one, it
allows callbacks like PyQt signal and slot connections to function
correctly.

This commit breaks the functionality of the Reset button in the python
console.