Add PythonQtPlugin files.
The PythonQt plugin makes the Qt API available to users from the
ParaView Python console.  The plugin uses the PythonQt library.  The
PythonQt homepage is pythonqt.sourceforge.net.

Note, PythonQt is different from PyQt4, it is a separate project and
uses a different technique to generate Qt bindings.

For more information, see the included README.

Change-Id: I05cd20422a4777866fbd1a5bc21783a6e06c2482